# docs/swagger.yaml
basePath: /
definitions:
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  handler.getAllUserModuleResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/tc.UserModule'
        type: array
    type: object
  handler.getAllUsersResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/tc.UsersList'
        type: array
    type: object
...
  handler.statusResponse:
    properties:
      status:
        type: string
    type: object
...
  tc.UpdateUserInput:
    properties:
      email:
        type: string
      id:
        type: string
      module_name:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    required:
    - email
    - name
    type: object
  tc.User:
    properties:
      email:
        type: string
      module_name:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    required:
    - email
    - name
    - password
    type: object
  tc.UserModule:
    properties:
      module_desc:
        type: string
      module_name:
        type: string
      user_id:
        type: string
    type: object
  tc.UsersList:
    properties:
      email:
        type: string
      id:
        type: string
      module_name:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
host: tc.kaztranscom.kz:9000
info:
  contact:
    email: 7773785631@mail.ru,K.Yatsenko@kaztranscom.kz
    name: Серышев А.В., Яценко К.Ю.
  description: API Сервер для Программы Total Control
  title: Total Control
  version: "1.0"
paths:
...
  /api/user/{id}:
    delete:
      consumes:
      - application/json
      description: Удаление Пользователя
      operationId: delete-user
      parameters:
      - description: uuid пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Удаление
      tags:
      - Пользователь
    get:
      consumes:
      - application/json
      description: Информация по пользователю по uuid пользователя
      operationId: get-user-by-id
      parameters:
      - description: uuid пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tc.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Информация
      tags:
      - Пользователь
    put:
      consumes:
      - application/json
      description: Редактирование информации о пользователе
      operationId: update-user
      parameters:
      - description: uuid пользователя
        in: path
        name: id
        required: true
        type: string
      - description: Поля для изменений
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/tc.UpdateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Редактирование
      tags:
      - Пользователь
  /api/user/list:
    get:
      consumes:
      - application/json
      description: Информация по всем пользователям
      operationId: get-users-list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllUsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Список всех пользователей
      tags:
      - Пользователь
  /api/user/module:
    get:
      consumes:
      - application/json
      description: Информация по модулям доступа авторизованного пользователя
      operationId: get-user-module
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllUserModuleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Информация по модулям
      tags:
      - Пользователь
  /api/user/module/{id}:
    get:
      consumes:
      - application/json
      description: Информация по модулям доступа пользователя, по uuid пользователя
      operationId: get-user-module-by-id
      parameters:
      - description: uuid пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllUserModuleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Информация по модулям
      tags:
      - Пользователь
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"